requirements.txt
django<1.9
pyaes==1.6.1
six==1.11.0
PyYAML==3.12
